requirements.txt
aioredis==1.2.0
alembic==1.0.0
amqp==2.3.2
aniso8601==3.0.2
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.1
Automat==0.7.0
bcrypt==3.1.4
beautifulsoup4==4.7.1
billiard==3.5.0.4
cachetools==3.1.0
certifi==2018.4.16
cffi==1.11.5
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
click==6.7
ConcurrentLogHandler==0.9.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.3.1
daphne==2.2.3
dj-database-url==0.5.0
Django==2.1.3
django-configurations==2.1
django-djconfig==0.8.0
django-graphos==0.3.41
django-haystack==2.8.1
django-infinite-scroll-pagination==0.2.0
django-spirit==0.7.0
django-timezone-field==3.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
dnspython==1.15.0
ephem==3.7.6.0
eventlet==0.24.1
Fabric3==1.14.post1
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-MySQL==1.4.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
funcsigs==1.0.2
get==2018.11.19
gevent==1.3.7
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
greenlet==0.4.15
gunicorn==19.9.0
hiredis==0.2.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
lxml==4.3.0
Mako==1.0.7
MarkupSafe==1.0
mistune==0.7.4
mock==2.0.0
monotonic==1.5
msgpack==0.5.6
nose2==0.7.4
olefile==0.46
paramiko==2.4.2
pbr==4.2.0
Pillow==5.3.0
post==2018.11.20
promise==2.2.1
psycopg2==2.7.5
psycopg2-binary==2.7.7
public==2018.11.20
pyasn1==0.4.4
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.6.4
PyMySQL==0.9.3
PyNaCl==1.3.0
python-crontab==2.3.5
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.5
query-string==2018.11.20
request==2018.11.20
requests==2.19.1
Rx==1.6.1
six==1.11.0
soupsieve==1.7.3
SQLAlchemy==1.2.10
Twisted==18.9.0
txaio==18.8.1
uni-slugify==0.1.4
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.2
Whoosh==2.7.4
zope.interface==4.6.0
